+++ lib/roadie/stylesheet.rb
@@ -21,18 +21,6 @@ module Roadie
       @blocks = parse_blocks(css.sub(BOM, ""))
     end
 
-    # @yield [selector, properties]
-    # @yieldparam [Selector] selector
-    # @yieldparam [Array<StyleProperty>] properties
-    # @deprecated Iterate over the #{blocks} instead. Will be removed on version 4.0.
-    def each_inlinable_block(&block)
-      # #map and then #each in order to support chained enumerations, etc. if
-      # no block is provided
-      inlinable_blocks.map { |style_block|
-        [style_block.selector, style_block.properties]
-      }.each(&block)
-    end
-
     def to_s
       blocks.join("\n")
     end